prefer explicit casting to `uint8_t` to drop higher bits

+++ include/challenge.hpp
@@ -40,7 +40,7 @@ chal_hash(std::span<const polynomial::polynomial_t, k> w,
       xof.absorb(_buffer);
     }
 
-    _buffer[buf_off] = w[coeff_idx / polynomial::N][coeff_idx % polynomial::N].raw();
+    _buffer[buf_off] = static_cast<uint8_t>(w[coeff_idx / polynomial::N][coeff_idx % polynomial::N].raw());
 
     buf_off++;
     coeff_idx++;
